ns: change PNP testing to use QNNPACK
Summary:

Changes the PNP test cases to use QNNPACK. The only reason is because
I'm switching to Mac M1 as my primary machine, which supports QNNPACK
but not fbgemm, and it's convenient for me to be able to run these
locally.

PNP itself is not backend specific, so it does not matter which backend
the functionality is tested on.

Test plan:

```
python test/test_quantization.py -k NShadows
```

ghstack-source-id: baf75dc3a2347bb67aa189513d4967ced49c75ad
Pull Request resolved: https://github.com/pytorch/pytorch/pull/91421